Move Archive.java to logic package
