+++ src/test/java/seedu/address/logic/archive/ArchiveTest.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.address.logic.archive;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -25,6 +25,8 @@ import seedu.address.model.Tracker;
 import seedu.address.model.module.Module;
 import seedu.address.model.module.ModuleCode;
 import seedu.address.model.module.ReadOnlyModule;
+import seedu.address.storage.JsonTrackerStorage;
+import seedu.address.storage.Storage;
 import seedu.address.testutil.ModelStub;
 import seedu.address.testutil.StorageStub;
 import seedu.address.testutil.TypicalModules;
